+++ classes/generator.py
@@ -1,6 +1,7 @@
 import importlib
 
-from classes.generator_error import GeneratorError
+from .generator_error import GeneratorError
+from .block import Block
 from .filter import Filter
 
 _HANDLERS_PATH = "handlers."
@@ -12,14 +13,14 @@ class Generator:
     Any processing required is up to the handler."""
 
     def generate(self, filter: Filter, output_filepath: str, handler_name: str, options: list[str]):
-        """Imports the module specified by handler and gives it a reference to the filter, a Block and a list of option strings.
+        """Imports the module specified by handler_name and gives it a reference to the filter, a Block and a list of option strings.
         The handler must be a python file under the directory "handlers", and the '.py' extension must be omitted.
         The handler must include the function handle(filter: Filter, block: Block, options: list[str])."""
         handler = self._get_handler(handler_name)
-        new_blocks = []
+        generated_raw_lines = []
         for block in filter.blocks:
-            new_blocks += handler.handle(filter, block, options)
-        return Filter(output_filepath, new_blocks)
+            generated_raw_lines += handler.handle(filter, block, options)
+        return Filter(output_filepath, Block.extract(generated_raw_lines))
 
     def _get_handler(self, handler_name: str):
         try: